Implement better buffer manager eviction policies.
The initial simplistic buffer manager has been replaced by
a more complex implementation.
Searching for assigned blocks happens in O(1) via a sync.Map.
Free buffers are maintained in a LIFO free list, and the manager
now uses a "clock hand" replacement policy.

To keep things relatively simple, the hand of the clock moves
randomly, following the iteration order of the sync.Map.
While this does not allow for deterministic eviction, it still
keeps the LRU flavour of the original algorithm.