Implement "copy" transaction operation.
Copy copies data from a location to another within the same page.
It is used for bulk operations where working with fixedlen or varlen
fields is not convenient.

The UNDO operation restores the previous bytes that have been
overwritten by the copy.

Along with the new operation, the new Compact method has been added
to the SlottedPage.
The method moves all non tombstoned tuples to the end of the free space,
effectively reclaiming space within the page.