simplify the API
The commit removes the Size type: it is redundant and differs
from the Offset type only in a nuanced conceptual way.
As a matter of facts, all reasonable and acceptable sizes
are expressed as offsets from the page buffers.

The commit also removes verbosity of unsafe methods.
These are all part of the storage package and it is understood
of their low level functioning.
The use of the package is meant to be part of the internals of
the system.
Unsafe operations are in this context an implementation detail
that does not need to be repeatedly highlighted in the function
names.